+++ pensa/__init__.py
@@ -7,7 +7,7 @@ from pyemma.util.contexts import settings
 import MDAnalysis as mda
 import matplotlib.pyplot as plt
 
-from .analysis import *
+from .features import *
 from .clusters import *
 from .comparison import *  
 from .pca import *
